requirements.txt
streamlit
transformers
sentence-transformers
faiss-cpu
PyPDF2
pytesseract
whisper
numpy
pillow
python-dotenv